Move is_checkmate to legal_moves module
