+++ tests/test_move_validation.py
@@ -3,71 +3,7 @@ import pytest
 from utahchess.board import Board
 from utahchess.move import REGULAR_MOVE, Move, make_move
 from utahchess.move_candidates import get_king_move_candidates, get_pawn_move_candidates
-from utahchess.move_validation import is_checkmate, is_valid_move
-
-
-def test_is_checkmate_fools_mate():
-    # when
-    board_string = f"""br-bn-bb-oo-bk-bb-bn-br
-            bp-bp-bp-bp-oo-bp-bp-bp
-            oo-oo-oo-oo-bp-oo-oo-oo
-            oo-oo-oo-oo-oo-oo-oo-oo
-            oo-oo-oo-oo-oo-oo-wp-bq
-            oo-oo-oo-oo-oo-wp-oo-oo
-            wp-wp-wp-wp-wp-oo-oo-wp
-            wr-wn-wb-wq-wk-wb-wn-wr"""
-    board = Board(board_string=board_string)
-
-    # then
-    assert is_checkmate(board=board, current_player="white")
-
-
-def test_is_checkmate_friendly_piece_can_save_king():
-    # when
-    board_string = f"""br-oo-wq-bq-bk-bb-oo-br
-            bp-bp-oo-oo-oo-oo-oo-oo
-            bb-oo-oo-oo-oo-bp-oo-oo
-            wp-oo-wb-oo-oo-oo-oo-wb
-            oo-oo-oo-oo-oo-oo-oo-oo
-            wn-oo-oo-oo-oo-oo-oo-oo
-            oo-wp-wk-oo-oo-oo-wp-oo
-            oo-wr-oo-bn-oo-oo-wn-wr"""
-    board = Board(board_string=board_string)
-
-    # then
-    assert not is_checkmate(board=board, current_player="black")
-
-
-def test_is_checkmate_friendly_piece_can_save_king_two():
-    # when
-    board_string = f"""br-oo-oo-oo-oo-bb-oo-br
-            bp-bp-oo-bb-oo-oo-bp-bp
-            bn-oo-oo-bp-oo-bk-oo-oo
-            oo-wb-oo-oo-bp-oo-oo-oo
-            oo-oo-wp-oo-wp-wp-bn-oo
-            wp-oo-wp-oo-oo-oo-oo-oo
-            oo-wb-oo-oo-oo-bq-wp-wp
-            wr-wn-oo-oo-oo-wq-wk-wr"""
-    board = Board(board_string=board_string)
-
-    # then
-    assert not is_checkmate(board=board, current_player="white")
-
-
-def test_is_checkmate_another_scenario():
-    # when
-    board_string = f"""br-oo-oo-wq-oo-bk-oo-br
-            oo-oo-oo-oo-bn-oo-bp-oo
-            oo-wb-oo-oo-wp-bp-oo-oo
-            oo-bp-oo-oo-oo-wp-wn-oo
-            bq-wp-oo-oo-wp-oo-oo-bp
-            bb-oo-wp-oo-oo-oo-oo-oo
-            oo-oo-oo-oo-oo-oo-oo-oo
-            oo-wn-oo-wb-oo-wn-oo-wr"""
-    board = Board(board_string=board_string)
-
-    # then
-    assert not is_checkmate(board=board, current_player="black")
+from utahchess.move_validation import is_valid_move
 
 
 def test_is_valid_move_restricted_king():
